Chore: move TopicContext to Contexts directory
