+++ client/Elements/Components/CreateTopic/CreateTopicWizard.patternfly.tsx
@@ -26,7 +26,7 @@ import './CreateTopicWizard.patternfly.css';
 import { CREATE_TOPIC } from 'Queries/Topics';
 import { useMutation } from '@apollo/client';
 import { NewTopic } from 'Entities/Entities.generated';
-import { CreateTopicAdvanceWizard } from '../CreateTopicAdvanceWizard/CreateTopicAdvanceWizard.patternfly';
+import { CreateTopicAdvanceWizard } from './CreateTopicAdvanceWizard.patternfly';
 
 interface ICreateTopicWizard {
   setIsCreateTopic: (value: boolean) => void;
