+++ client/Elements/Components/CreateTopic/utils.ts
@@ -7,12 +7,12 @@ const capitalizeText = (text: string) => {
   return text.charAt(0).toUpperCase() + text.slice(1);
 };
 
-export const kebabToCamel = (kebabCase: string) => {
-  const words: string[] = kebabCase.split("-");
+export const kebabToCamel = function (kebabCase: string): string {
+  const words: string[] = kebabCase.split('-');
 
   const camelCased: string[] = words.map((word: string, index: number) => {
-    return index > 0? capitalizeText(word) : word;
+    return index > 0 ? capitalizeText(word) : word;
   });
 
-  return camelCased.join("");
+  return camelCased.join('');
 };
